requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
html2text>=2024.2
playwright>=1.40
orjson>=3.9
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0
click>=8.1
python-dotenv>=1.0
yt-dlp>=2024.1
youtube-transcript-api>=1.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.9
langdetect>=1.0.9
PyMuPDF>=1.23.0
# NLP & OCR features
easyocr>=1.7
underthesea>=6.7
scikit-learn>=1.3
rake-nltk>=1.0.6
# Anti-bot stealth
playwright-stealth>=0.1
fake-useragent>=1.4
# CrawlKit v0.1.0
